Aggregate tests by a grouping variable

Counts test results by a grouping variable (Created Date, by default) and adds some useful columns

Usage

aggregate_tests(data, by = "Created Date")

Arguments

data

A data.frame() or tibble().

by

A string matching one of the column names of data; the default is "Created Date".

Value

an object of the same type as data, with the following columns added:

n tests

n with results

n positive

% positive

period

The output will be arranged in ascending order of the variable identified by the by argument; that is, it will be arranged in chronological order of Created Date, by default.